Cleanup includes

+++ src/utils/enum-array.h
@@ -12,20 +12,20 @@ struct EnumArray
 {
   using IndexType = t_Index;
   using ElementType = t_Element;
-  static u32 const size = (u32)IndexType::COUNT - (u32)IndexType::START;
+  static size_t const size = (size_t)IndexType::COUNT - (size_t)IndexType::START;
 
   ElementType const&
   operator[](IndexType index) const
   {
     assert(index >= IndexType::START && index < IndexType::COUNT);
-    u32 offset = (u32)index - (u32)IndexType::START;
+    size_t offset = (size_t)index - (size_t)IndexType::START;
     return members[offset];
   }
   ElementType&
   operator[](IndexType index)
   {
     assert(index >= IndexType::START && index < IndexType::COUNT);
-    u32 offset = (u32)index - (u32)IndexType::START;
+    size_t offset = (size_t)index - (size_t)IndexType::START;
     return members[offset];
   }
